+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug when loading menu on mobile
+
 ### Changed
 
 - Changed Link paths in Submenu, fixed some translations